generalized C listener code
